database_service/requirements.txt
anyio==3.5.0; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.5.0; python_version >= "3.7"
click==8.0.3; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
environs==9.5.0; python_version >= "3.6"
fastapi==0.71.0; python_full_version >= "3.6.1"
h11==0.13.0; python_version >= "3.7"
idna==3.3; python_version >= "3.6" and python_full_version >= "3.6.2"
loguru==0.5.3; python_version >= "3.5"
marshmallow==3.14.1; python_version >= "3.6"
pydantic==1.9.0; python_full_version >= "3.6.1"
pymongo==4.0.1; python_version >= "3.6"
python-dotenv==0.19.2; python_version >= "3.6"
python-multipart==0.0.5
pyyaml==6.0; python_version >= "3.6"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.17.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.1.0; python_version >= "3.6" and python_full_version >= "3.6.1"
uvicorn==0.17.4; python_version >= "3.7"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
